Separate ICES controller
